requirements.txt
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.3
django-allauth==0.38.0
django-crispy-forms==1.7.2
django-elasticsearch-dsl==6.4.2
django-haystack==2.8.1
django-leaflet==0.24.0
django-multiselectfield==0.1.8
django-photologue==3.8.1
django-sortedm2m==1.5.0
djangorestframework==3.10.2
elasticsearch==6.4.0
elasticsearch-dsl==6.4.0
ExifRead==2.1.2
idna==2.7
oauthlib==2.1.0
Pillow==5.3.0
psycopg2==2.7.6
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
six==1.12.0
urllib3==1.24.1